pyproject.toml
[tool.poetry]
name = "verthandi"
version = "0.1.0"
description = "CLI time-tracking utility"
authors = ["Jeppe Klitgaard <jeppe@dapj.dk>"]
license = "MIT"

    [tool.poetry.scripts]
    verthandi = "verthandi.cli.main:app"
    v = "verthandi.cli.main:app"

    [tool.poetry.dependencies]
    python = "^3.8"
    requests = "^2.24.0"
    typer = "^0.3.2"
    appdirs = "^1.4.4"
    tomlkit = "^0.7.0"
    orjson = "^3.3.1"
    pydantic = "^1.6.1"
    rich = "^7.0.0"
    python-dateutil = "^2.8.1"

    [tool.poetry.dev-dependencies]
    pylint = "^2.6.0"
    black = "^20.8b1"
    isort = "^5.5.3"
    mypy = "^0.782"
    invoke = "^1.4.1"
    pytest = "^6.0.2"

[tool.pylint]
    [tool.pylint.MASTER]
    extension-pkg-whitelist = [
        "orjson",
    ]

    [tool.pylint.FORMAT]
    max-line-length = 88
    good-names = [
        "i", "j", "k", "x", "y", "z", "f", "c", "v", "e", "d", "r", "t", "p",
        "db", "id", "fd", "df",
        "op", "eq", "gt", "lt",
        "s3", "pk", "up",
    ]

# Code compliance
[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
combine_as_imports = true

known_api = ["requests"]
known_first_party = ["verthandi"]

sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "API",
    "FIRSTPARTY",
    "LOCALFOLDER"
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
